feat: add pagination proto, move search results to widgets
