+++ src/features/pagination/ui/Pagination.module.scss
@@ -1,0 +1,5 @@
+.pagination {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--space-s);
+}
