+++ src/widgets/index.ts
@@ -1,0 +1,2 @@
+export { Header } from './header';
+export { SearchResults } from './search-results';
