+++ src/widgets/search-results/index.ts
@@ -1,0 +1,1 @@
+export { SearchResults } from './ui/SearchResults';
