+++ src/widgets/search-results/ui/SearchResults.module.scss
@@ -1,0 +1,3 @@
+.pagination {
+  align-self: center;
+}
